code/requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8.5,<3.0
djangorestframework>=3.11,<4.0
gunicorn>=20.1.0,<21.0.0
whitenoise>=5.2.0,<6.0.0
py-sudoku
imutils==0.5.4
opencv-python==4.5.2.54
scikit-image==0.18.2
tensorflow==2.5.0